pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-code-review"
version = "0.1.0"
description = "AI Code Review CLI with Gemini"
authors = [
    { name = "Author", email = "author@example.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "google-generativeai>=0.3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
aicr = "aicr.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["aicr"]
